CEAS_website/requirements.txt
Django==1.10.2
django-multiselectfield==0.1.4
et-xmlfile==1.0.1
jdcal==1.3
openpyxl==2.4.1
psycopg2==2.6.2